requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
AC-Flask-Hipchat==0.2.10
click==6.6                # via flask
enum34==1.1.6
Flask==0.11.1             # via ac-flask-hipchat
itsdangerous==0.24        # via flask
Jinja2==2.8               # via flask
MarkupSafe==0.23          # via jinja2
PyJWT==1.4.2              # via ac-flask-hipchat
pymongo==3.3.0            # via ac-flask-hipchat
redis==2.10.5             # via ac-flask-hipchat
requests==2.11.1          # via ac-flask-hipchat
Werkzeug==0.11.11         # via flask
